Phase 2-5: Fix MyPy errors and improve test infrastructure
- Fixed type errors in debug_utils (null count iteration, Callable import)
- Fixed dict entry type incompatibility in validators
- Added return type annotations to exception __str__ methods
- Improved mock_db_connection with UTF-8 encoding support
- Fixed mock_minio_client to return iterables instead of subscriptables
- Reduced test failures from 44 to 10 (76% improvement)
- MyPy errors: 52 (some new strict checks enabled)

+++ src/data_quality/validators.py
@@ -332,9 +332,9 @@ class DataQualityValidator:
                         {
                             "expectation_type": "expect_column_values_to_not_be_null",
                             "column": col,
-                            "unexpected_count": null_count,
-                            "unexpected_percent": (1.0 - completeness) * 100,
-                            "kwargs": {"mostly": completeness_threshold},
+                            "unexpected_count": int(null_count),
+                            "unexpected_percent": float((1.0 - completeness) * 100),
+                            "kwargs": {"mostly": float(completeness_threshold)},
                         }
                     )
